WIP
WIP

adding model for finish test

replaced alert with modal and moved some code to base page

pre-commit before moving code to base page

moving to basepage 1

moving to basepage 2

move code to base page 4

move code to base page 5

update css of finish test modal

removed cyclic dependancy

updated references in office page b from page state to use common references

format import
