+++ src/app/pages/office/cat-b/office.cat-b.page.ts
@@ -3,62 +3,22 @@ import {
   Platform,
   ToastController,
   AlertController,
+  ModalController,
 } from '@ionic/angular';
 import { Component } from '@angular/core';
-import { PracticeableBasePageComponent } from '@shared/classes/practiceable-base-page';
 import { AuthenticationProvider } from '@providers/authentication/authentication';
 import { Store, select } from '@ngrx/store';
 import { StoreModel } from '@shared/models/store.model';
 import { getTests } from '@store/tests/tests.reducer';
-import { getTestSummary } from '@store/tests/test-summary/test-summary.reducer';
-import { merge, Observable, Subscription } from 'rxjs';
-import { FormGroup } from '@angular/forms';
+import { Observable } from 'rxjs';
 import {
   getCurrentTest,
   getTestOutcome,
-  isTestOutcomeSet,
-  isPassed,
-  getTestOutcomeText,
-  getActivityCode,
-  getJournalData,
 } from '@store/tests/tests.selector';
-import {
-  getRouteNumber,
-  getCandidateDescription,
-  getAdditionalInformation,
-  getWeatherConditions,
-  getIdentification,
-  getIndependentDriving,
-} from '@store/tests/test-summary/test-summary.selector';
 import { map, withLatestFrom } from 'rxjs/operators';
-import {
-  WeatherConditions,
-  Identification,
-  IndependentDriving,
-} from '@dvsa/mes-test-schema/categories/common';
-import { startsWith } from 'lodash';
 import { TestCategory } from '@dvsa/mes-test-schema/category-definitions/common/test-category';
 import { CatBUniqueTypes } from '@dvsa/mes-test-schema/categories/B';
-import {
-  RouteNumberChanged,
-  IndependentDrivingTypeChanged,
-  IdentificationUsedChanged,
-  CandidateDescriptionChanged,
-  WeatherConditionsChanged,
-  AdditionalInformationChanged,
-} from '@store/tests/test-summary/test-summary.actions';
-import { getCandidate } from '@store/tests/journal-data/common/candidate/candidate.reducer';
-import {
-  getCandidateName,
-  getCandidateDriverNumber,
-  formatDriverNumber,
-} from '@store/tests/journal-data/common/candidate/candidate.selector';
 import { QuestionProvider } from '@providers/question/question';
-import {
-  getTestSlotAttributes,
-} from '@store/tests/journal-data/common/test-slot-attributes/test-slot-attributes.reducer';
-import { getTestDate, getTestStartDateTime, getTestTime }
-  from '@store/tests/journal-data/common/test-slot-attributes/test-slot-attributes.selector';
 import {
   getETA,
   getETAFaultText,
@@ -66,7 +26,6 @@ import {
   getEcoFaultText,
   getShowMeQuestionOptions,
 } from '@store/tests/test-data/common/test-data.selector';
-// TODO: update import for category specific page version
 import {
   getVehicleChecks,
   getSelectedTellMeQuestionText,
@@ -74,77 +33,35 @@ import {
   getTellMeQuestion,
 } from '@store/tests/test-data/cat-b/test-data.cat-b.selector';
 import { getTestData } from '@store/tests/test-data/cat-b/test-data.reducer';
-import { PersistTests } from '@store/tests/tests.actions';
-import { WeatherConditionSelection } from '@providers/weather-conditions/weather-conditions.model';
 import { WeatherConditionProvider } from '@providers/weather-conditions/weather-condition';
-import {
-  AddDangerousFaultComment,
-} from '@store/tests/test-data/common/dangerous-faults/dangerous-faults.actions';
-import { AddSeriousFaultComment } from '@store/tests/test-data/common/serious-faults/serious-faults.actions';
-import { AddDrivingFaultComment } from '@store/tests/test-data/common/driving-faults/driving-faults.actions';
-import {
-  ShowMeQuestionSelected,
-  AddShowMeTellMeComment,
-} from '@store/tests/test-data/cat-b/vehicle-checks/vehicle-checks.actions';
-import {
-  AddControlledStopComment,
-} from '@store/tests/test-data/common/controlled-stop/controlled-stop.actions';
-import { AddManoeuvreComment } from '@store/tests/test-data/common/manoeuvres/manoeuvres.actions';
-import { EyesightTestAddComment } from '@store/tests/test-data/common/eyesight-test/eyesight-test.actions';
 import { CommentSource, FaultSummary } from '@shared/models/fault-marking.model';
 import { OutcomeBehaviourMapProvider } from '@providers/outcome-behaviour-map/outcome-behaviour-map';
-import { ActivityCodeModel, activityCodeModelList } from '@shared/constants/activity-code/activity-code.constants';
-import { CompetencyOutcome } from '@shared/models/competency-outcome';
-import { getRekeyIndicator } from '@store/tests/rekey/rekey.reducer';
-import { isRekey } from '@store/tests/rekey/rekey.selector';
-import { SetActivityCode } from '@store/tests/activity-code/activity-code.actions';
+import { ActivityCodeModel } from '@shared/constants/activity-code/activity-code.constants';
 import { VehicleChecksQuestion } from '@providers/question/vehicle-checks-question.model';
 import { FaultCountProvider } from '@providers/fault-count/fault-count';
 import { getTestCategory } from '@store/tests/category/category.reducer';
 import { FaultSummaryProvider } from '@providers/fault-summary/fault-summary';
-import { getNewTestStartTime } from '@shared/helpers/test-start-time';
-import { SetStartDate }
-  from '@store/tests/journal-data/common/test-slot-attributes/test-slot-attributes.actions';
 import { Router } from '@angular/router';
+import { OfficeBasePageComponent } from '@shared/classes/test-flow-base-pages/office/office-base-page';
+import { AddDangerousFaultComment } from '@store/tests/test-data/common/dangerous-faults/dangerous-faults.actions';
+import { startsWith } from 'lodash';
+import { AddManoeuvreComment } from '@store/tests/test-data/common/manoeuvres/manoeuvres.actions';
+import { CompetencyOutcome } from '@shared/models/competency-outcome';
+import { AddControlledStopComment } from '@store/tests/test-data/common/controlled-stop/controlled-stop.actions';
 import {
-  OfficeViewDidEnter,
-  CompleteTest,
-  SavingWriteUpForLater,
-  OfficeValidationError,
-  TestStartDateChanged,
-} from '../office.actions';
-import { JOURNAL_PAGE, TestFlowPageNames } from '../../page-names.constants';
-import { behaviourMap } from '../office-behaviour-map';
+  AddShowMeTellMeComment,
+  ShowMeQuestionSelected,
+} from '@store/tests/test-data/cat-b/vehicle-checks/vehicle-checks.actions';
+import { AddSeriousFaultComment } from '@store/tests/test-data/common/serious-faults/serious-faults.actions';
+import { EyesightTestAddComment } from '@store/tests/test-data/common/eyesight-test/eyesight-test.actions';
+import { AddDrivingFaultComment } from '@store/tests/test-data/common/driving-faults/driving-faults.actions';
 
 interface OfficePageState {
-  activityCode$: Observable<ActivityCodeModel>;
-  startTime$: Observable<string>;
-  startDate$: Observable<string>;
-  startDateTime$: Observable<string>;
-  testOutcome$: Observable<string>;
-  testOutcomeText$: Observable<string>;
-  isPassed$: Observable<boolean>;
-  isTestOutcomeSet$: Observable<boolean>;
-  candidateName$: Observable<string>;
-  candidateDriverNumber$: Observable<string>;
-  routeNumber$: Observable<number>;
-  displayRouteNumber$: Observable<boolean>;
-  displayIndependentDriving$: Observable<boolean>;
-  displayCandidateDescription$: Observable<boolean>;
-  displayIdentification$: Observable<boolean>;
   displayShowMeQuestion$: Observable<boolean>;
   displayTellMeQuestion$: Observable<boolean>;
-  displayWeatherConditions$: Observable<boolean>;
-  displayAdditionalInformation$: Observable<boolean>;
-  displayEco$: Observable<boolean>;
-  displayEta$: Observable<boolean>;
   displayDrivingFault$: Observable<boolean>;
   displaySeriousFault$: Observable<boolean>;
   displayDangerousFault$: Observable<boolean>;
-  identification$: Observable<Identification>;
-  independentDriving$: Observable<IndependentDriving>;
-  candidateDescription$: Observable<string>;
-  additionalInformation$: Observable<string>;
   showMeQuestion$: Observable<VehicleChecksQuestion>;
   showMeQuestionOptions$: Observable<VehicleChecksQuestion[]>;
   tellMeQuestionText$: Observable<string>;
@@ -153,10 +70,8 @@ interface OfficePageState {
   drivingFaults$: Observable<FaultSummary[]>;
   drivingFaultCount$: Observable<number>;
   displayDrivingFaultComments$: Observable<boolean>;
-  weatherConditions$: Observable<WeatherConditions[]>;
   dangerousFaults$: Observable<FaultSummary[]>;
   seriousFaults$: Observable<FaultSummary[]>;
-  isRekey$: Observable<boolean>;
 }
 
 @Component({
@@ -164,50 +79,50 @@ interface OfficePageState {
   templateUrl: 'office.cat-b.page.html',
   styleUrls: ['../office.page.scss'],
 })
-export class OfficeCatBPage extends PracticeableBasePageComponent {
+export class OfficeCatBPage extends OfficeBasePageComponent {
+
   pageState: OfficePageState;
-  form: FormGroup;
-  toast: any;
   drivingFaultCtrl: string = 'drivingFaultCtrl';
   seriousFaultCtrl: string = 'seriousFaultCtrl';
   dangerousFaultCtrl: string = 'dangerousFaultCtrl';
   static readonly maxFaultCount = 15;
-  subscription: Subscription;
 
-  weatherConditions: WeatherConditionSelection[];
   showMeQuestions: VehicleChecksQuestion[];
   activityCodeOptions: ActivityCodeModel[];
-  startDateTime: string;
-  isValidStartDateTime: boolean = true;
 
   constructor(
+    platform: Platform,
+    authenticationProvider: AuthenticationProvider,
+    router: Router,
     store$: Store<StoreModel>,
-    public toastController: ToastController,
-    public navController: NavController,
-    public platform: Platform,
-    public authenticationProvider: AuthenticationProvider,
-    private weatherConditionProvider: WeatherConditionProvider,
+    navController: NavController,
+    toastController: ToastController,
+    modalController: ModalController,
+    outcomeBehaviourProvider: OutcomeBehaviourMapProvider,
+    weatherConditionProvider: WeatherConditionProvider,
     public questionProvider: QuestionProvider,
-    private outcomeBehaviourProvider: OutcomeBehaviourMapProvider,
     public alertController: AlertController,
     private faultCountProvider: FaultCountProvider,
     private faultSummaryProvider: FaultSummaryProvider,
-    public router: Router,
   ) {
-    super(platform, authenticationProvider, router, store$);
-    this.form = new FormGroup({});
-    this.weatherConditions = this.weatherConditionProvider.getWeatherConditions();
-    this.showMeQuestions = questionProvider.getShowMeQuestions(TestCategory.B);
-    this.outcomeBehaviourProvider.setBehaviourMap(behaviourMap);
-    this.activityCodeOptions = activityCodeModelList;
-  }
+    super(
+      platform,
+      authenticationProvider,
+      router,
+      store$,
+      navController,
+      toastController,
+      modalController,
+      outcomeBehaviourProvider,
+      weatherConditionProvider,
+    );
 
-  ionViewDidEnter(): void {
-    this.store$.dispatch(OfficeViewDidEnter());
+    this.showMeQuestions = questionProvider.getShowMeQuestions(TestCategory.B);
+
   }
 
   ngOnInit(): void {
-    super.ngOnInit();
+    super.onInitialisation();
     const currentTest$ = this.store$.pipe(
       select(getTests),
       select(getCurrentTest),
@@ -216,90 +131,7 @@ export class OfficeCatBPage extends PracticeableBasePageComponent {
       select(getTestCategory),
     );
     this.pageState = {
-      activityCode$: currentTest$.pipe(
-        select(getActivityCode),
-      ),
-      isRekey$: currentTest$.pipe(
-        select(getRekeyIndicator),
-        select(isRekey),
-      ),
-      testOutcome$: currentTest$.pipe(
-        select(getTestOutcome),
-      ),
-      testOutcomeText$: currentTest$.pipe(
-        select(getTestOutcomeText),
-      ),
-      isPassed$: currentTest$.pipe(
-        select(isPassed),
-      ),
-      isTestOutcomeSet$: currentTest$.pipe(
-        select(isTestOutcomeSet),
-      ),
-      startTime$: currentTest$.pipe(
-        select(getJournalData),
-        select(getTestSlotAttributes),
-        select(getTestTime),
-      ),
-      startDate$: currentTest$.pipe(
-        select(getJournalData),
-        select(getTestSlotAttributes),
-        select(getTestDate),
-      ),
-      startDateTime$: currentTest$.pipe(
-        select(getJournalData),
-        select(getTestSlotAttributes),
-        select(getTestStartDateTime),
-      ),
-      candidateName$: currentTest$.pipe(
-        select(getJournalData),
-        select(getCandidate),
-        select(getCandidateName),
-      ),
-      candidateDriverNumber$: currentTest$.pipe(
-        select(getJournalData),
-        select(getCandidate),
-        select(getCandidateDriverNumber),
-        map(formatDriverNumber),
-      ),
-      routeNumber$: currentTest$.pipe(
-        select(getTestSummary),
-        select(getRouteNumber),
-      ),
-      displayRouteNumber$: currentTest$.pipe(
-        select(getTestOutcome),
-        withLatestFrom(currentTest$.pipe(
-          select(getTestSummary),
-          select(getRouteNumber),
-        )),
-        map(([outcome, route]) => this.outcomeBehaviourProvider.isVisible(outcome, 'routeNumber', route)),
-      ),
-      displayIndependentDriving$: currentTest$.pipe(
-        select(getTestOutcome),
-        withLatestFrom(currentTest$.pipe(
-          select(getTestSummary),
-          select(getIndependentDriving),
-        )),
-        map(([outcome, independent]) =>
-          this.outcomeBehaviourProvider.isVisible(outcome, 'independentDriving', independent)),
-      ),
-      displayCandidateDescription$: currentTest$.pipe(
-        select(getTestOutcome),
-        withLatestFrom(currentTest$.pipe(
-          select(getTestSummary),
-          select(getCandidateDescription),
-        )),
-        map(([outcome, candidate]) =>
-          this.outcomeBehaviourProvider.isVisible(outcome, 'candidateDescription', candidate)),
-      ),
-      displayIdentification$: currentTest$.pipe(
-        select(getTestOutcome),
-        withLatestFrom(currentTest$.pipe(
-          select(getTestSummary),
-          select(getIdentification),
-        )),
-        map(([outcome, identification]) =>
-          this.outcomeBehaviourProvider.isVisible(outcome, 'identification', identification)),
-      ),
+      ...this.commonPageState,
       displayShowMeQuestion$: currentTest$.pipe(
         select(getTestOutcome),
         withLatestFrom(currentTest$.pipe(
@@ -319,42 +151,6 @@ export class OfficeCatBPage extends PracticeableBasePageComponent {
         )),
         map(([outcome, question]) =>
           this.outcomeBehaviourProvider.isVisible(outcome, 'tellMeQuestion', question)),
-      ),
-      displayWeatherConditions$: currentTest$.pipe(
-        select(getTestOutcome),
-        withLatestFrom(currentTest$.pipe(
-          select(getTestSummary),
-          select(getWeatherConditions),
-        )),
-        map(([outcome, weather]) =>
-          this.outcomeBehaviourProvider.isVisible(outcome, 'weatherConditions', weather)),
-      ),
-      displayAdditionalInformation$: currentTest$.pipe(
-        select(getTestOutcome),
-        withLatestFrom(currentTest$.pipe(
-          select(getTestSummary),
-          select(getAdditionalInformation),
-        )),
-        map(([outcome, additional]) =>
-          this.outcomeBehaviourProvider.isVisible(outcome, 'additionalInformation', additional)),
-      ),
-      displayEco$: currentTest$.pipe(
-        select(getTestOutcome),
-        withLatestFrom(currentTest$.pipe(
-          select(getTestData),
-          select(getEco),
-        )),
-        map(([outcome, eco]) =>
-          this.outcomeBehaviourProvider.isVisible(outcome, 'eco', eco)),
-      ),
-      displayEta$: currentTest$.pipe(
-        select(getTestOutcome),
-        withLatestFrom(currentTest$.pipe(
-          select(getTestData),
-          select(getETA),
-        )),
-        map(([outcome, eta]) =>
-          this.outcomeBehaviourProvider.isVisible(outcome, 'eta', eta)),
       ),
       displayDrivingFault$: currentTest$.pipe(
         select(getTestOutcome),
@@ -392,22 +188,6 @@ export class OfficeCatBPage extends PracticeableBasePageComponent {
             this.faultSummaryProvider.getDangerousFaultsList(testData, TestCategory.B),
           )),
       ),
-      candidateDescription$: currentTest$.pipe(
-        select(getTestSummary),
-        select(getCandidateDescription),
-      ),
-      independentDriving$: currentTest$.pipe(
-        select(getTestSummary),
-        select(getIndependentDriving),
-      ),
-      identification$: currentTest$.pipe(
-        select(getTestSummary),
-        select(getIdentification),
-      ),
-      additionalInformation$: currentTest$.pipe(
-        select(getTestSummary),
-        select(getAdditionalInformation),
-      ),
       showMeQuestion$: currentTest$.pipe(
         select(getTestData),
         select(getVehicleChecks),
@@ -415,7 +195,8 @@ export class OfficeCatBPage extends PracticeableBasePageComponent {
       ),
       showMeQuestionOptions$: currentTest$.pipe(
         select(getTestOutcome),
-        map((outcome) => getShowMeQuestionOptions(this.showMeQuestions, outcome, this.outcomeBehaviourProvider)),
+        map((outcome) =>
+          getShowMeQuestionOptions(this.showMeQuestions, outcome, this.outcomeBehaviourProvider)),
       ),
       tellMeQuestionText$: currentTest$.pipe(
         select(getTestData),
@@ -434,15 +215,21 @@ export class OfficeCatBPage extends PracticeableBasePageComponent {
       ),
       dangerousFaults$: currentTest$.pipe(
         select(getTestData),
-        map((data) => this.faultSummaryProvider.getDangerousFaultsList(data, TestCategory.B)),
+        withLatestFrom(category$),
+        map(([testData, category]) =>
+          this.faultSummaryProvider.getDangerousFaultsList(testData, category as TestCategory)),
       ),
       seriousFaults$: currentTest$.pipe(
         select(getTestData),
-        map((data) => this.faultSummaryProvider.getSeriousFaultsList(data, TestCategory.B)),
+        withLatestFrom(category$),
+        map(([testData, category]) =>
+          this.faultSummaryProvider.getSeriousFaultsList(testData, category as TestCategory)),
       ),
       drivingFaults$: currentTest$.pipe(
         select(getTestData),
-        map((data) => this.faultSummaryProvider.getDrivingFaultsList(data, TestCategory.B)),
+        withLatestFrom(category$),
+        map(([testData, category]) =>
+          this.faultSummaryProvider.getDrivingFaultsList(testData, category as TestCategory)),
       ),
       drivingFaultCount$: currentTest$.pipe(
         select(getTestData),
@@ -455,86 +242,13 @@ export class OfficeCatBPage extends PracticeableBasePageComponent {
         select(getTestData),
         map((data) => this.shouldDisplayDrivingFaultComments(data)),
       ),
-      weatherConditions$: currentTest$.pipe(
-        select(getTestSummary),
-        select(getWeatherConditions),
-      ),
     };
 
     this.setupSubscriptions();
   }
 
-  setupSubscriptions() {
-    const {
-      startDateTime$,
-    } = this.pageState;
-    this.subscription = merge(
-      startDateTime$.pipe(map((value) => this.startDateTime = value)),
-    ).subscribe();
-  }
-
-  ionViewDidLeave(): void {
-    if (this.subscription) {
-      this.subscription.unsubscribe();
-    }
-  }
-
-  async popToRoot() {
-    if (this.isPracticeMode) {
-      this.exitPracticeMode();
-      return;
-    }
-    await this.navController.navigateBack(JOURNAL_PAGE);
-  }
-
-  defer() {
-    this.popToRoot();
-    this.store$.dispatch(SavingWriteUpForLater());
-    this.store$.dispatch(PersistTests());
-  }
-
-  onSubmit() {
-    if (this.isFormValid()) {
-      this.showFinishTestModal();
-    }
-  }
-
-  setIsValidStartDateTime(isValid: boolean) {
-    this.isValidStartDateTime = isValid;
-  }
-
-  dateOfTestChanged(inputDate: string) {
-    const customStartDate = getNewTestStartTime(inputDate, this.startDateTime);
-    this.store$.dispatch(TestStartDateChanged(this.startDateTime, customStartDate));
-    this.store$.dispatch(SetStartDate(customStartDate));
-  }
-
   showMeQuestionChanged(showMeQuestion: VehicleChecksQuestion): void {
     this.store$.dispatch(ShowMeQuestionSelected(showMeQuestion));
-  }
-
-  identificationChanged(identification: Identification): void {
-    this.store$.dispatch(IdentificationUsedChanged(identification));
-  }
-
-  independentDrivingChanged(independentDriving: IndependentDriving): void {
-    this.store$.dispatch(IndependentDrivingTypeChanged(independentDriving));
-  }
-
-  weatherConditionsChanged(weatherConditions: WeatherConditions[]): void {
-    this.store$.dispatch(WeatherConditionsChanged(weatherConditions));
-  }
-
-  routeNumberChanged(routeNumber: number) {
-    this.store$.dispatch(RouteNumberChanged(routeNumber));
-  }
-
-  candidateDescriptionChanged(candidateDescription: string) {
-    this.store$.dispatch(CandidateDescriptionChanged(candidateDescription));
-  }
-
-  additionalInformationChanged(additionalInformation: string): void {
-    this.store$.dispatch(AdditionalInformationChanged(additionalInformation));
   }
 
   dangerousFaultCommentChanged(dangerousFaultComment: FaultSummary) {
@@ -615,75 +329,6 @@ export class OfficeCatBPage extends PracticeableBasePageComponent {
       this.store$.dispatch(AddShowMeTellMeComment(drivingFaultComment.comment));
     }
 
-  }
-
-  activityCodeChanged(activityCodeModel: ActivityCodeModel) {
-    const { showMeQuestion } = this.form.controls;
-    if (showMeQuestion) {
-      if (showMeQuestion.value && showMeQuestion.value.code === 'N/A') {
-        this.form.controls['showMeQuestion'].setValue({});
-      }
-    }
-    this.store$.dispatch(SetActivityCode(activityCodeModel.activityCode));
-  }
-
-  private createToast = async (errorMessage: string) => {
-    this.toast = await this.toastController.create({
-      message: errorMessage,
-      position: 'top',
-      cssClass: 'mes-toast-message-error',
-      duration: 5000,
-      buttons: [{
-        text: 'X',
-        role: 'cancel',
-      }],
-    });
-  };
-
-  async showFinishTestModal() {
-    const alert = await this.alertController.create({
-      header: 'Are you sure you wish to mark the write up for this test as complete?',
-      cssClass: 'finish-test-modal',
-      buttons: [
-        {
-          text: 'Back',
-          handler: () => { },
-        },
-        {
-          text: 'Continue',
-          handler: () => this.completeTest(),
-        },
-      ],
-    });
-    await alert.present();
-  }
-
-  async goToReasonForRekey() {
-    if (await this.isFormValid()) {
-      await this.router.navigate([TestFlowPageNames.REKEY_REASON_PAGE]);
-    }
-  }
-
-  async isFormValid() {
-    Object.keys(this.form.controls).forEach((controlName) => this.form.controls[controlName].markAsDirty());
-    if (this.form.valid) {
-      return true;
-    }
-    Object.keys(this.form.controls).forEach((controlName) => {
-      if (this.form.controls[controlName].invalid) {
-        this.store$.dispatch(OfficeValidationError(`${controlName} is blank`));
-      }
-    });
-    await this.createToast('Fill all mandatory fields');
-    this.toast.present();
-    return false;
-  }
-
-  completeTest() {
-    if (!this.isPracticeMode) {
-      this.store$.dispatch(CompleteTest());
-    }
-    this.popToRoot();
   }
 
   shouldDisplayDrivingFaultComments = (data: CatBUniqueTypes.TestData): boolean => {
