+++ src/app/pages/office/components/fault-comment/fault-comment.scss
@@ -1,0 +1,5 @@
+:host {
+  .text-area-padding {
+    padding-left: var(--ion-padding, 16px);
+  }
+}
