+++ src/app/pages/office/components/fault-comment/fault-comment.ts
@@ -18,6 +18,7 @@ enum ValidFaultTypes {
 @Component({
   selector: 'fault-comment',
   templateUrl: 'fault-comment.html',
+  styleUrls: ['fault-comment.scss'],
 })
 
 export class FaultCommentComponent implements OnChanges {
