+++ src/app/pages/office/components/finish-test-modal/finish-test-modal.html
@@ -1,0 +1,28 @@
+<ion-card class="modal-alert">
+  <modal-alert-title id="model-title">Mark as complete?</modal-alert-title>
+  <ion-row>
+    <ion-col class="ion-no-padding" size="15"></ion-col>
+    <ion-col class="ion-no-padding" size="66">
+      <div class="modal-alert-message">
+        <div id='model-text' class="mes-data">Are your sure you wish to mark the write up for this test as complete?</div>
+      </div>
+    </ion-col>
+    <ion-col class="ion-no-padding" size="15"></ion-col>
+  </ion-row>
+  <ion-row>
+    <ion-col class="column-border-top ion-no-padding">
+      <ion-row>
+        <ion-col>
+          <ion-button id="cancel-button" fill="clear" class="finish-test-cancel-button" (click)="onBack()">
+            Back
+          </ion-button>
+        </ion-col>
+        <ion-col>
+          <ion-button id="continue-button" fill="clear" class="finish-test-continue-button" (click)="onCompleteTest()">
+            Continue
+          </ion-button>
+        </ion-col>
+      </ion-row>
+    </ion-col>
+  </ion-row>
+</ion-card>
