+++ src/app/pages/office/components/finish-test-modal/finish-test-modal.scss
@@ -1,0 +1,27 @@
+@import "src/theme/sass-partials/modal";
+
+:host {
+  @extend .modal-alert-wrapper;
+
+  .modal-alert {
+    width: 400px;
+  }
+
+  .modal-alert-message {
+    margin-bottom: 24px;
+  }
+
+  .mes-align-content-center {
+    text-align: center;
+  }
+
+  .finish-test-cancel-button {
+    color: var(--gds-black);
+    border-right: 1px solid var(--gds-grey-3);
+  }
+
+  .finish-test-continue-button {
+    color: var(--gds-black);
+    font-weight: bold;
+  }
+}
