+++ src/app/pages/office/components/finish-test-modal/finish-test-modal.ts
@@ -1,0 +1,31 @@
+import {
+  ModalController,
+} from '@ionic/angular';
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'finish-test-modal',
+  templateUrl: './finish-test-modal.html',
+  styleUrls: ['./finish-test-modal.scss'],
+})
+export class FinishTestModal {
+
+  completeTest: Function;
+
+  constructor(
+    public modalController: ModalController,
+  ) {
+  }
+
+  ngOnInit(): void {
+  }
+
+  onCompleteTest = async () => {
+    await this.modalController.dismiss();
+    await this.completeTest;
+  };
+
+  onBack = async () => {
+    await this.modalController.dismiss();
+  };
+}
