+++ src/app/pages/office/components/route-number/route-number.html
@@ -7,8 +7,12 @@
     <ion-row class="spacing-row"></ion-row>
     <ion-row>
       <ion-col size="12">
-        <input type="number" id="route" class="vehicle-reg-input mes-data" numbersOnly
-        formControlName="routeNumber" (change)="routeNumberChanged($event.target.value)">
+        <input type="number"
+               id="route"
+               class="vehicle-reg-input input-styling-background mes-data"
+               numbersOnly
+               formControlName="routeNumber"
+               (change)="routeNumberChanged($event.target.value)">
       </ion-col>
     </ion-row>
     <ion-row class="validation-message-row ion-align-items-center">
