--- src/app/pages/office/office.model.ts
@@ -1,5 +1,0 @@
-export type OfficeModel = {
-  dangerousFaultComments: string[],
-  drivingFaultComments: string[],
-  seriousFaultComments: string[],
-};
