+++ src/app/pages/office/office.page.scss
@@ -1,160 +1,177 @@
-:host::ng-deep {
+:host ::ng-deep {
   .driving-fault-no-commentary-row {
     padding: 0;
     padding-bottom: 22px;
+
     .col {
       padding: 0;
     }
+
     .component-label {
       padding-left: 16px;
     }
   }
+
   .mes-primary-button {
     min-width: 150px;
   }
+
   .character-count {
     color: var(--gds-grey-1);
     font-weight: initial;
     letter-spacing: initial;
+
     &.ng-invalid {
       font-weight: 600;
       letter-spacing: -0.49px;
     }
   }
+
   ion-content {
-        ion-select {
-          max-width: initial;
-          width: 100%;
-          margin-right: 48px;
-        }
-        textarea {
-          width: 100%;
-        }
-        input.vehicle-reg-input {
-          width: 66px;
-        }
-        .mes-full-width-card {
-          height: 120px;
-          #continue-to-test-report {
-            width: 100%;
-            height: 56px;
-          }
-        }
-        .manual-label-top-spacing {
-          label {
-            display: block;
-            margin-top: 42px;
-          }
-        }
-        .fault-badge.col {
-          padding-left: 16px;
-        }
-        .component-label.col {
-          padding-left: 16px;
-        }
-        ion-col[align-self-start].component-label {
-          padding-top: 46px;
-        }
-        .textarea-input.col {
-          padding-right: 48px;
-        }
-        .check-row {
-          padding-top: 4px;
-        }
-        .icon-image {
-          width: 25px;
-          height: 25px;
-          min-width: 25px;
-        }
-      }
-      #test-outcome-text {
-        .pass {
-          color: var(--gds-dark-green-3);
-        }
-        .fail {
-          color: var(gds-red);
-        }
-        .terminated {
-          color: var(--gds-grey-1);
-        }
-      }
-      .inline-mode {
-        display: inline;
-      }
-      .card-content-ios{
-        padding: 0;
-      }
-      .grid{
-        padding: 0;
-      }
-      .align-col-right {
-        float: right;
-        text-align: right;
-      }
-    #defer-button {
-      --background: var(--gds-grey-3);
-      --box-shadow: 0px 2px 0px 0px rgba(0, 0, 0, 0.5);
-      --border-radius: 10px;
-      margin: 0 0.2rem;
-      padding: 0 1.6em;
-      &.activated {
-        --background: var(--gds-grey-4);
-      }
-      h3 {
-        --color: var(--mes-black);
-        .text-zoom-regular &,
-        .text-zoom-large &,
-        .text-zoom-x-large & {
-          line-height: 2em;
-        }
-        min-height: 44px;
-        padding-top: 1px
+    ion-select {
+      max-width: initial;
+      width: 100%;
+      margin-right: 48px;
+    }
+
+    textarea {
+      width: 100%;
+    }
+
+    input.vehicle-reg-input {
+      width: 66px;
+    }
+
+    .mes-full-width-card {
+      height: 120px;
+
+      #continue-to-test-report {
+        width: 100%;
+        height: 56px;
       }
     }
-    .fault-heading {
-      padding-left: 20px;
+
+    .manual-label-top-spacing {
+      label {
+        display: block;
+        margin-top: 42px;
+      }
     }
-    .fault-label {
-      margin-right: 16px;
+
+    .fault-badge.col {
+      padding-left: 16px;
     }
-    .keyboard-space {
-      margin-bottom: 220px;
+
+    .component-label.col {
+      padding-left: 16px;
     }
+
+    ion-col[align-self-start].component-label {
+      padding-top: 46px;
+    }
+
+    .textarea-input.col {
+      padding-right: 48px;
+    }
+
+    .check-row {
+      padding-top: 4px;
+    }
+
+    .icon-image {
+      width: 25px;
+      height: 25px;
+      min-width: 25px;
+    }
+  }
+
+  #test-outcome-text {
+    .pass {
+      color: var(--gds-dark-green-3);
+    }
+
+    .fail {
+      color: var(gds-red);
+    }
+
+    .terminated {
+      color: var(--gds-grey-1);
+    }
+  }
+
+  .inline-mode {
+    display: inline;
+  }
+
+  .card-content-ios {
+    padding: 0;
+  }
+
+  .grid {
+    padding: 0;
+  }
+
+  .align-col-right {
+    float: right;
+    text-align: right;
+  }
+
+  #defer-button {
+    --background: var(--gds-grey-3);
+    --box-shadow: 0px 2px 0px 0px rgba(0, 0, 0, 0.5);
+    --border-radius: 10px;
+    margin: 0 0.2rem;
+    padding: 0 1.6em;
+
+    &.activated {
+      --background: var(--gds-grey-4);
+    }
+
+    h3 {
+      --color: var(--mes-black);
+
+      .text-zoom-regular &,
+      .text-zoom-large &,
+      .text-zoom-x-large & {
+        line-height: 2em;
+      }
+
+      min-height: 44px;
+      padding-top: 1px
+    }
+  }
+
+  .fault-heading {
+    padding-left: 20px;
+  }
+
+  .fault-label {
+    margin-right: 16px;
+  }
+
+  .keyboard-space {
+    margin-bottom: 220px;
+  }
+
   ion-footer {
     ion-row.mes-full-width-card {
       padding: 24px 32px;
       background-color: var(--mes-white);
+
       button {
         width: initial;
         min-height: 56px;
       }
+
       .full-width {
         width: 100%;
       }
     }
   }
-}
 
-
-.finish-test-modal {
-  .alert-head:before {
-    display: inline-block;
-    font-size: 68px;
-    content: '\f276';
-    display: inline-block;
-    font-family: "Ionicons";
-    -moz-osx-font-smoothing: grayscale;
-    -webkit-font-smoothing: antialiased;
-    font-style: normal;
-    font-variant: normal;
-    font-weight: normal;
-    line-height: 1;
-    text-rendering: auto;
-    text-transform: none;
+  .office-button {
+    width: 300px;
+    min-height: 56px;
   }
-  .alert-title {
-    padding-left: 72px;
-    padding-right: 72px;
-  }
-}
 
+}
