--- src/app/pages/office/office.reducer.ts
@@ -1,16 +1,0 @@
-import * as officeActions from './office.actions';
-import { OfficeModel } from './office.model';
-
-export const initialState: OfficeModel = {
-  dangerousFaultComments: [],
-  drivingFaultComments: [],
-  seriousFaultComments: [],
-};
-
-export function officeReducer(state = initialState, action: officeActions.OfficeActionTypes): OfficeModel {
-  switch (action.type) {
-    default:
-      return state;
-  }
-
-}
