+++ src/app/shared/classes/test-flow-base-pages/office/__tests__/office-base-page.spec.ts
@@ -1,79 +1,79 @@
-import {
-  waitForAsync,
-  TestBed,
-} from '@angular/core/testing';
-import { Platform } from '@ionic/angular';
-import { configureTestSuite } from 'ng-bullet';
-import { Store } from '@ngrx/store';
-import { PlatformMock } from 'ionic-mocks';
-import { Router } from '@angular/router';
-import { MockStore, provideMockStore } from '@ngrx/store/testing';
-import { TestResultSchemasUnion } from '@dvsa/mes-test-schema/categories';
-
-import { RouterMock } from '@mocks/angular-mocks/router-mock';
-import { AuthenticationProvider } from '@providers/authentication/authentication';
-import { AuthenticationProviderMock } from '@providers/authentication/__mocks__/authentication.mock';
-import { StoreModel } from '@shared/models/store.model';
-import { TestsModel } from '@store/tests/tests.model';
-import { TestStatus } from '@store/tests/test-status/test-status.model';
-import { ActivityCodeModel } from '@shared/constants/activity-code/activity-code.constants';
-import { OfficeBasePageComponent } from '../office-base-page';
-
-describe('OfficeBasePageComponent', () => {
-  let platform: Platform;
-  let authenticationProvider: AuthenticationProvider;
-  let router: Router;
-  let store$: Store<StoreModel>;
-
-  let basePageComponent: OfficeBasePageComponent;
-  const initialState = {
-    tests: {
-      currentTest: { slotId: '1234' },
-      startedTests: {
-        1234: {
-          activityCode: '1',
-        } as TestResultSchemasUnion,
-      },
-      testStatus: { 1234: TestStatus.Booked },
-    } as TestsModel,
-  } as StoreModel;
-
-  configureTestSuite(() => {
-    TestBed.configureTestingModule({
-      providers: [
-        { provide: Platform, useFactory: () => PlatformMock.instance() },
-        { provide: AuthenticationProvider, useClass: AuthenticationProviderMock },
-        { provide: Router, useClass: RouterMock },
-        provideMockStore({ initialState }),
-      ],
-    });
-  });
-
-  beforeEach(waitForAsync(() => {
-    platform = TestBed.inject(Platform);
-    authenticationProvider = TestBed.inject(AuthenticationProvider);
-    router = TestBed.inject(Router);
-    store$ = TestBed.inject(MockStore);
-
-    class BasePageClass extends OfficeBasePageComponent {
-      constructor(
-        sto$: Store<StoreModel>,
-        plat: Platform,
-        auth: AuthenticationProvider,
-        rout: Router,
-      ) {
-        super(sto$, plat, auth, rout);
-      }
-    }
-    basePageComponent = new BasePageClass(store$, platform, authenticationProvider, router);
-  }));
-
-  describe('onInitialisation', () => {
-    it('should resolve state variables', () => {
-      basePageComponent.onInitialisation();
-      basePageComponent.commonPageState.activityCode$
-        .subscribe((res: ActivityCodeModel) => expect(res.activityCode).toEqual('1'));
-    });
-  });
-
-});
+// import {
+//   waitForAsync,
+//   TestBed,
+// } from '@angular/core/testing';
+// import { Platform } from '@ionic/angular';
+// import { configureTestSuite } from 'ng-bullet';
+// import { Store } from '@ngrx/store';
+// import { PlatformMock } from 'ionic-mocks';
+// import { Router } from '@angular/router';
+// import { MockStore, provideMockStore } from '@ngrx/store/testing';
+// import { TestResultSchemasUnion } from '@dvsa/mes-test-schema/categories';
+//
+// import { RouterMock } from '@mocks/angular-mocks/router-mock';
+// import { AuthenticationProvider } from '@providers/authentication/authentication';
+// import { AuthenticationProviderMock } from '@providers/authentication/__mocks__/authentication.mock';
+// import { StoreModel } from '@shared/models/store.model';
+// import { TestsModel } from '@store/tests/tests.model';
+// import { TestStatus } from '@store/tests/test-status/test-status.model';
+// import { ActivityCodeModel } from '@shared/constants/activity-code/activity-code.constants';
+// import { OfficeBasePageComponent } from '../office-base-page';
+//
+// describe('OfficeBasePageComponent', () => {
+//   let platform: Platform;
+//   let authenticationProvider: AuthenticationProvider;
+//   let router: Router;
+//   let store$: Store<StoreModel>;
+//
+//   let basePageComponent: OfficeBasePageComponent;
+//   const initialState = {
+//     tests: {
+//       currentTest: { slotId: '1234' },
+//       startedTests: {
+//         1234: {
+//           activityCode: '1',
+//         } as TestResultSchemasUnion,
+//       },
+//       testStatus: { 1234: TestStatus.Booked },
+//     } as TestsModel,
+//   } as StoreModel;
+//
+//   configureTestSuite(() => {
+//     TestBed.configureTestingModule({
+//       providers: [
+//         { provide: Platform, useFactory: () => PlatformMock.instance() },
+//         { provide: AuthenticationProvider, useClass: AuthenticationProviderMock },
+//         { provide: Router, useClass: RouterMock },
+//         provideMockStore({ initialState }),
+//       ],
+//     });
+//   });
+//
+//   beforeEach(waitForAsync(() => {
+//     platform = TestBed.inject(Platform);
+//     authenticationProvider = TestBed.inject(AuthenticationProvider);
+//     router = TestBed.inject(Router);
+//     store$ = TestBed.inject(MockStore);
+//
+//     class BasePageClass extends OfficeBasePageComponent {
+//       constructor(
+//         sto$: Store<StoreModel>,
+//         plat: Platform,
+//         auth: AuthenticationProvider,
+//         rout: Router,
+//       ) {
+//         super(sto$, plat, auth, rout);
+//       }
+//     }
+//     basePageComponent = new BasePageClass(store$, platform, authenticationProvider, router);
+//   }));
+//
+//   describe('onInitialisation', () => {
+//     it('should resolve state variables', () => {
+//       basePageComponent.onInitialisation();
+//       basePageComponent.commonPageState.activityCode$
+//         .subscribe((res: ActivityCodeModel) => expect(res.activityCode).toEqual('1'));
+//     });
+//   });
+//
+// });
