+++ src/components/common/activity-code/activity-code.html
@@ -3,7 +3,7 @@
     <ion-col size="32" class="component-label align-vertical-centre">
         <label>Activity code</label>
     </ion-col>
-    <ion-col size="63">
+    <ion-col size="63" class="ion-align-self-center ion-padding-start">
         <ion-row class="spacing-row"></ion-row>
         <ion-row>
             <ion-select id="activity-code-selector"
