+++ src/theme/sass-partials/_alert.scss
@@ -65,7 +65,7 @@
       position: absolute;
       margin: 0;
       right: 10px;
-      background-color: var(--gds-grey-5);
+      background-color: var(--gds-grey-3);
       border: none;
     }
     $alert-checkbox-button-padding: 4px 0;
@@ -83,8 +83,9 @@
         div.alert-checkbox-inner {
           width: 13px;
           height: 22px;
-          margin-left: 5px;
+          margin-left: 4px;
           border-width: 0 5px 5px 0;
+          background-color: var(--gds-dark-green);
         }
       }
       div.alert-checkbox-label {
